% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attributes.R
\name{add_attributes}
\alias{add_attributes}
\title{Add attributes to a list of nodes}
\usage{
add_attributes(input_list)
}
\arguments{
\item{input_list}{}
}
\description{
Add attributes to a list of nodes
}
