% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_check_executable_status.R
\name{add_check_executable_status}
\alias{add_check_executable_status}
\title{Checks each actions executable state based on available data}
\usage{
add_check_executable_status(actions, ui_data, data_context)
}
\arguments{
\item{actions}{A data.table containing information on dependencies, output columns, actions etc.
\code{can_execute}, \code{missing_dependencies}, and \code{lineage}}

\item{data_context}{An R6 object of class \link{data_context} that contains
information on available data in a given \code{connector} object}
}
\value{
A data.table containing action data with the additional columns
}
\description{
Checks each actions executable state based on available data
}
