% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_check_executable_status.R
\name{add_default_execution_status}
\alias{add_default_execution_status}
\title{Add default execution status when no data context is provided}
\usage{
add_default_execution_status(actions)
}
\description{
Add default execution status when no data context is provided
}
