% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_keys_to_depend_cols.R
\name{add_foreign_key_as_depends_col}
\alias{add_foreign_key_as_depends_col}
\title{Add Foreign Key Columns to Dependency Columns for External Data Domains}
\usage{
add_foreign_key_as_depends_col(x, index_echo, nodes_echo_external, pk)
}
\arguments{
\item{x}{A \code{data.table} representing the ADaM nodes structure.}

\item{index_echo}{An integer vector of indices corresponding to col_echo
actions in the nodes.}

\item{nodes_echo_external}{A logical vector of the same length as
\code{index_echo} indicating whether an action references external dependencies.}

\item{pk}{A named list containing primary key column details for external
domains.}
}
\value{
A modified \code{data.table} with enriched foreign key dependency columns
for external actions.
}
\description{
Adds foreign key columns to \code{depend_cols} in nodes for actions
with external dependencies.
}
\details{
Iterates through the specified indices of col_echo actions and
enriches their dependency columns by appending domain-specific foreign key
information. Stops execution if an unrecognized domain is encountered.
}
