% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_implied_predecessors.R
\name{add_implied_predecessors}
\alias{add_implied_predecessors}
\title{Add predecessors not specified by the user}
\usage{
add_implied_predecessors(nodes)
}
\value{
the node data model with new rows for the implied predecessors. These
new rows have type == "implied_predecessor"
}
\description{
".self" columns that are consumed by derivation nodes need to be
listed as the outputs of other nodes. If there is a column that is not
listed as an output of another node at this stage, it is assumed to be an
"implied predecessor", i.e. a predecessor that was not specified by the
user, but required.

Implied predecessors can also include "temporary variable", aka columns
that are used for derivations, but are not part of the final dataset
}
