% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_initialize_domain_nodes.R
\name{add_initialize_domain_nodes}
\alias{add_initialize_domain_nodes}
\title{Add initialize_domain nodes}
\usage{
add_initialize_domain_nodes(program_order, nodes_6)
}
\arguments{
\item{nodes_6}{}
}
\description{
Add initialize_domain nodes
}
