% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nodes_to_load_external_data.R
\name{add_nodes_to_load_external_data}
\alias{add_nodes_to_load_external_data}
\title{Update program order with nodes to load external data}
\usage{
add_nodes_to_load_external_data(program_order, nodes)
}
\arguments{
\item{nodes}{}
}
\description{
Update program order with nodes to load external data
}
