% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nodes_to_load_external_data.R
\name{add_nodes_to_read_data}
\alias{add_nodes_to_read_data}
\title{Update program order with nodes to load external data}
\usage{
add_nodes_to_read_data(program_order, nodes, init_metadata, domain_keys)
}
\arguments{
\item{domain_keys}{}
}
\description{
Update program order with nodes to load external data
}
