% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_program_init_nodes.R
\name{add_program_init_nodes}
\alias{add_program_init_nodes}
\title{Add program init nodes}
\usage{
add_program_init_nodes(program_order, nodes_6)
}
\arguments{
\item{nodes_6}{}
}
\description{
Add program init nodes
}
