% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_read_data_nodes.R
\name{add_read_data_nodes}
\alias{add_read_data_nodes}
\title{Update program order with nodes to load external data}
\usage{
add_read_data_nodes(program_order, nodes, init_metadata, domain_keys)
}
\arguments{
\item{domain_keys}{}
}
\description{
Update program order with nodes to load external data
}
