% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_program_init_nodes.R
\name{add_read_domain_nodes}
\alias{add_read_domain_nodes}
\title{Add read_domain nodes}
\usage{
add_read_domain_nodes(program_order, nodes_6)
}
\arguments{
\item{nodes_6}{}
}
\description{
Add read_domain nodes
}
