% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_actions.R
\name{apply_priority_1}
\alias{apply_priority_1}
\title{Apply Priority 1: Select nodes from last selected node's domain}
\usage{
apply_priority_1(candidate_nodes, sorted_nodes, lkp_dom, primary_domain)
}
\description{
Apply Priority 1: Select nodes from last selected node's domain
}
