% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_consolidated_env.R
\name{assert_all_code_id_sources_exist}
\alias{assert_all_code_id_sources_exist}
\title{Assert All Code ID Sources Exist}
\usage{
assert_all_code_id_sources_exist(x, code_ids)
}
\arguments{
\item{x}{List containing function sources divided into \code{pkg} and \code{file}
entries, each containing details of loaded functions with names as keys.}

\item{code_ids}{Character vector containing function names that should be
present in the sources.}
}
\value{
Returns TRUE invisibly if all code IDs are found, otherwise raises
an error with details about missing functions.
}
\description{
Verifies that all specified code IDs are found in the loaded
function sources, raising an error if any are missing.
}
