% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_consolidated_env.R
\name{assert_source_available}
\alias{assert_source_available}
\title{Assert Source Availability}
\usage{
assert_source_available(item, check_type = c("package", "file"))
}
\arguments{
\item{item}{Character string specifying the name of the package or file to
check.}

\item{check_type}{Character string, either "package" or "file", indicating
the type of source to check. Defaults to "package".}
}
\value{
Returns TRUE invisibly if the specified package is available or the
file exists. Otherwise raises an error
}
\description{
Verifies the availability of specified packages or files,
raising an error if they are not available.
}
\details{
Checks whether a given package is installed and available, or whether
a specified file exists. An error is raised with details if the resource is
not available.
}
