% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_valid_depend_cols.R
\name{assert_valid_depend_cols}
\alias{assert_valid_depend_cols}
\title{Validate Dependency Columns}
\usage{
assert_valid_depend_cols(
  actions,
  ui_yml,
  domain_keys,
  check_cross_domain_adam_dependencies
)
}
\arguments{
\item{ui_yml}{A list containing UI configuration, which includes the
initialization parameters needed for filtering dependency columns.}

\item{check_cross_domain_adam_dependencies}{}

\item{x}{A data.frame containing action metadata, including columns for
'type', 'domain', 'depend_cols', and 'outputs'.}
}
\value{
An invisible copy of the input data.frame if all dependencies are valid.
}
\description{
This function checks if all specified dependency columns for each action
in a given dataset are present in the corresponding outputs. If any dependency
columns are missing, it generates an error message indicating the missing columns
and the actions they are required for.
}
