% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_valid_ui_yml.R
\name{assert_valid_domain_spec}
\alias{assert_valid_domain_spec}
\title{Validate Individual Domain Specification}
\usage{
assert_valid_domain_spec(domain_spec, domain_name)
}
\arguments{
\item{domain_spec}{The domain specification to validate}

\item{domain_name}{The name of the domain (for error messages)}
}
\value{
Invisible TRUE if valid, otherwise stops with error
}
\description{
Validates the structure of a single domain specification
}
