% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_valid_ui_yml.R
\name{assert_valid_init_spec}
\alias{assert_valid_init_spec}
\title{Validate Init Specification}
\usage{
assert_valid_init_spec(init_spec, domain_name)
}
\arguments{
\item{init_spec}{The init specification to validate}

\item{domain_name}{The domain name (for error messages)}
}
\value{
Invisible TRUE if valid, otherwise stops with error
}
\description{
Validates the structure of the init section
}
