% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_valid_trial_config.R
\name{assert_valid_trial_config}
\alias{assert_valid_trial_config}
\title{Assert trial_configuration is valid}
\usage{
assert_valid_trial_config(trial_metadata)
}
\arguments{
\item{trial_metadata}{A list containing trial metadata.}
}
\value{
Nothing if trial configuration contains all needed fields
}
\description{
Validates that the trial configuration is a list and contains
the required fields.
}
