% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_valid_adam_dependencies.R
\name{check_missing_dependency}
\alias{check_missing_dependency}
\title{Check if Dependency is Missing}
\usage{
check_missing_dependency(y, missing_deps)
}
\arguments{
\item{y}{Data frame or list containing 'domain' and 'column_name' columns}

\item{missing_deps}{Character vector of missing dependencies}
}
\value{
Logical value indicating if any dependencies are missing
}
\description{
Determines if any column dependencies are in the missing dependencies list.
}
\details{
Checks if any fully qualified column names (domain.column_name)
from the dependency list are found in the missing dependencies list.
}
