% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_external_data_domains.R
\name{classify_external_data_domains_2}
\alias{classify_external_data_domains_2}
\title{Classify external data domains}
\usage{
classify_external_data_domains_2(vector)
}
\arguments{
\item{vector}{}
}
\description{
Classify external data domains
}
