% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_origin_type_columns.R
\name{collapse_origin_type_columns}
\alias{collapse_origin_type_columns}
\title{Collapse origin into Type column}
\usage{
collapse_origin_type_columns(nodes)
}
\arguments{
\item{nodes}{}
}
\value{
a data.table with the information from \code{origin} moved to \code{type} and
\code{origin} deleted
}
\description{
We don't need an origin column, all the information can be
contained in a single column "type", and this simplifies downstream
processing. We don't treat all "column" nodes the same anyways, we need to
distinguish between pred and derived anyways.
}
