% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_primary_keys.R
\name{collate_primary_keys}
\alias{collate_primary_keys}
\title{Collate primary keys}
\usage{
collate_primary_keys(trial_metadata_keys)
}
\value{
a list where each element is a named vector conaing the priarmy keys
for that domain
}
\description{
Collects the primary key information produces a single list
}
