% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_edges.R
\name{combine_and_clean_edges}
\alias{combine_and_clean_edges}
\title{Combine and clean all edges}
\usage{
combine_and_clean_edges(column_edges, row_edges)
}
\arguments{
\item{column_edges}{Edges from column dependencies}

\item{row_edges}{Edges from row dependencies}
}
\value{
A cleaned data.table of unique edges
}
\description{
Combines column and row dependency edges, removes
self-references and duplicates
}
