% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_node_list_to_dt.R
\name{convert_node_list_to_dt}
\alias{convert_node_list_to_dt}
\title{Convert node list to data.table}
\usage{
convert_node_list_to_dt(nodes)
}
\arguments{
\item{nodes}{}
}
\description{
Convert node list to data.table
}
