% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_edges.R
\name{create_column_dependency_edges}
\alias{create_column_dependency_edges}
\title{Create edges based on column dependencies}
\usage{
create_column_dependency_edges(nodes)
}
\arguments{
\item{nodes}{The nodes data.table}
}
\value{
A data.table of edges based on column dependencies
}
\description{
Identifies parent-child relationships based on column
dependencies
}
