% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_consolidated_env.R
\name{create_consolidated_env}
\alias{create_consolidated_env}
\title{Create consolidate environment for all code components}
\usage{
create_consolidated_env(packages = NULL, source_files = NULL, code_ids)
}
\arguments{
\item{source_files}{}
}
\value{
Returns an environment
}
\description{
Create consolidate environment for all code components
}
