% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_domain_initialize_nodes.R
\name{create_domain_initialize_nodes}
\alias{create_domain_initialize_nodes}
\title{Replace "core" dependencies with domain initialization nodes}
\usage{
create_domain_initialize_nodes(nodes, core_vars, domain_init_data)
}
\arguments{
\item{domain_init_data}{}
}
\description{
Adds new nodes representing the domain initialization steps, and
removes the corresponding nodes that represent the SDTM core variables
}
