% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_check_executable_status.R
\name{create_lineage_message}
\alias{create_lineage_message}
\title{Create lineage message for missing dependencies}
\usage{
create_lineage_message(
  message,
  action,
  dependencies = NULL,
  report_missing_outputs = FALSE
)
}
\description{
Create lineage message for missing dependencies
}
