% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_preprocess_domain_nodes.R
\name{create_preprocess_domain_nodes}
\alias{create_preprocess_domain_nodes}
\title{Replace "core" dependencies with domain initialization nodes}
\usage{
create_preprocess_domain_nodes(nodes)
}
\arguments{
\item{nodes}{}
}
\description{
Adds new nodes representing the domain initialization steps, and
removes the corresponding nodes that represent the SDTM core variables
}
