% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_edges.R
\name{create_row_dependency_edges}
\alias{create_row_dependency_edges}
\title{Create edges based on row dependencies}
\usage{
create_row_dependency_edges(nodes)
}
\arguments{
\item{nodes}{The nodes data.table}
}
\value{
A data.table of edges based on row dependencies, or NULL if none
exist
}
\description{
Identifies parent-child relationships based on explicit row
dependencies
}
