% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_depend_cols.R
\name{enrich_core_compute_actions}
\alias{enrich_core_compute_actions}
\title{Enrich Core Column Dependencies for col_compute Actions}
\usage{
enrich_core_compute_actions(x)
}
\arguments{
\item{x}{A \code{data.table} of nodes representing ADaM domain metadata.}
}
\value{
A modified \code{data.table} of nodes with enriched \code{depend_cols} for
col_compute actions.
}
\description{
Enriches \code{depend_cols} for col_compute actions involving core
columns.
}
\details{
Identifies col_compute actions that share core column dependencies
and ensures the proper execution order by appending the outputs of other
relevant actions to their dependency columns.

Steps:
\itemize{
\item Detect core columns used as dependencies and returned as outputs.
\item For each col_compute action, check if other actions within the same domain share
the same core dependencies.
\item Append the output columns of these related actions to the \code{depend_cols} of the
col_compute action being processed.
}
}
