% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_with_external_dependencies.R
\name{enrich_with_external_dependencies}
\alias{enrich_with_external_dependencies}
\title{Parses nodes for external dependencies, including "core" variables}
\usage{
enrich_with_external_dependencies(nodes, init_metadata)
}
\arguments{
\item{init_metadata}{}
}
\value{
a list object where the name of the list element corresponds to the program_id in \code{grouped_nodes}
}
\description{
Finds all "external" dependencies for each "node"
}
