% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_edges.R
\name{expand_child_columns}
\alias{expand_child_columns}
\title{Expand child columns for each node}
\usage{
expand_child_columns(nodes)
}
\arguments{
\item{nodes}{The nodes data.table}
}
\value{
A data.table with expanded child column information
}
\description{
Creates a detailed list of columns each node produces
}
