% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_edges.R
\name{expand_parent_columns}
\alias{expand_parent_columns}
\title{Expand parent columns for each node}
\usage{
expand_parent_columns(nodes)
}
\arguments{
\item{nodes}{The nodes data.table}
}
\value{
A data.table with expanded parent column information
}
\description{
Creates a detailed list of columns each node depends on
}
