% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_read_data_nodes.R
\name{external_dependencies_per_program}
\alias{external_dependencies_per_program}
\title{Get external dependencies per program}
\usage{
external_dependencies_per_program(
  program_order,
  nodes,
  init_metadata,
  domain_keys
)
}
\arguments{
\item{init_metadata}{}
}
\description{
Get external dependencies per program
}
