% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sdtm_core_variables.R
\name{extract_sdtm_core_variables}
\alias{extract_sdtm_core_variables}
\title{Title}
\usage{
extract_sdtm_core_variables(nodes)
}
\arguments{
\item{nodes}{}
}
\description{
We want all SDTM core variables. This is determined as the
variables that are the parents of "predecessor" nodes. But only predecessor
nodes that have a single parent
}
