% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_actions.R
\name{find_remaining_upstream_dependencies}
\alias{find_remaining_upstream_dependencies}
\title{Find domains that still have remaining upstream dependencies}
\usage{
find_remaining_upstream_dependencies(adj_matrix, lkp_dom)
}
\description{
Find domains that still have remaining upstream dependencies
}
