% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_top_two_levels.R
\name{flatten_top_two_levels}
\alias{flatten_top_two_levels}
\title{Flatten top two level}
\usage{
flatten_top_two_levels(input_list)
}
\arguments{
\item{input_list}{}
}
\description{
Used when re-shaping the metadata model so that it can be
efficently processed to make the dependency graph. Without this, it would
require much more traversing of the list structure
}
