% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ui_data.R
\name{format_code_id}
\alias{format_code_id}
\title{Format code ID for node ID}
\usage{
format_code_id(code_id)
}
\arguments{
\item{code_id}{Vector of code IDs to format}
}
\value{
Vector of formatted code ID strings
}
\description{
Formats code IDs for inclusion in node IDs.
}
\details{
Prepends a hyphen to non-NA code IDs, or returns an empty string for NA values.
}
