% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ui_data.R
\name{format_parameters}
\alias{format_parameters}
\title{Format parameters for node ID}
\usage{
format_parameters(parameters)
}
\arguments{
\item{parameters}{Vector of parameters to format}
}
\value{
Vector of formatted parameter strings
}
\description{
Formats parameters for inclusion in node IDs.
}
\details{
Prepends a hyphen to non-NA parameters, or returns an empty string for NA values.
}
