% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_action_code.R
\name{generate_action_code}
\alias{generate_action_code}
\title{Generate Code for Individual Nodes}
\usage{
generate_action_code(
  actions_program_i,
  domain_keys,
  code_component_envr,
  ui_data,
  trial_metadata,
  path_output
)
}
\arguments{
\item{actions_program_i}{Data table containing nodes for a specific program
with columns for type, domain, outputs, dependencies, and parameters}

\item{domain_keys}{Named list mapping domain names to their respective key
columns}

\item{ui_data}{List with domain-specific configurations and filter
dependencies}

\item{trial_metadata}{List containing trial-specific metadata including data
paths and connection information
study}

\item{path_output}{Character string specifying the output path where
generated data should be stored}

\item{std_code_env}{Environment containing code components}
}
\value{
A list containing generated code blocks for each node in the
program, with each element representing the code for one processing step
}
\description{
Generates executable code for each node within a program based
on node type and specifications.
}
