% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adam_code.R
\name{generate_adam_code}
\alias{generate_adam_code}
\title{Generates the complete set of ADaM programs}
\usage{
generate_adam_code(
  path_ui_data,
  path_std_lib,
  path_trial_metadata,
  path_domain_keys,
  path_output,
  data_connection = c("connector", "pharmaverse"),
  trial_metadata_path
)
}
\arguments{
\item{data_connection}{}
}
\description{
Generates the complete set of ADaM programs
}
