% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_init_domain_code.R
\name{generate_init_domain_code}
\alias{generate_init_domain_code}
\title{Write the R code to initialize ADaM domain in "initial" programs}
\usage{
generate_init_domain_code(.self, keep_columns, source_domains)
}
\arguments{
\item{domain_filters_exist}{}
}
\description{
Write the R code to initialize ADaM domain in "initial" programs
}
