% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_initialize_domain.R
\name{generate_initialize_domain}
\alias{generate_initialize_domain}
\title{Write the R code to initialize ADaM domain in "initial" programs}
\usage{
generate_initialize_domain(.self, core_domains, domain_filters_exist)
}
\arguments{
\item{domain_filters_exist}{}
}
\description{
Write the R code to initialize ADaM domain in "initial" programs
}
