% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_preprocess_domain.R
\name{generate_preprocess_domain}
\alias{generate_preprocess_domain}
\title{Write the R code to initialize a domain}
\usage{
generate_preprocess_domain(
  .self,
  base_domains,
  adsl_domain_keys,
  filter_domain = NULL,
  filter_global = NULL,
  filter_depend_cols = NULL,
  keep_vars = NULL
)
}
\arguments{
\item{.self}{Character of length 1. Name of domain}

\item{base_domains}{list of core SDTM domains used to make "base" table}

\item{filter_global}{String. Global filter to be applied after merging}

\item{keep_vars}{Vector of strings naming variables from SDTM core domains
kept}
}
\description{
Write the R code to initialize a domain
}
