% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_program.R
\name{generate_program}
\alias{generate_program}
\title{Generate program(s)}
\usage{
generate_program(
  program_order,
  nodes,
  domain_keys,
  code_component_env,
  trial_metadata,
  ui_data,
  data_connection,
  path_output = NULL
)
}
\arguments{
\item{data_connection}{}

\item{path_output}{When data_connection is "pharmaverse" the generated
programs need to point to a location where the outputs are stored}
}
\description{
Generate program(s)
}
