% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_read_domain.R
\name{generate_read_domain}
\alias{generate_read_domain}
\title{Write the R code to read ADaM domain in "update" program}
\usage{
generate_read_domain(adam_domain)
}
\arguments{
\item{data_connection}{}
}
\description{
Write the R code to read ADaM domain in "update" program
}
