% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_business_logic.R
\name{get_business_rules}
\alias{get_business_rules}
\title{Get business rules for a given schema}
\usage{
get_business_rules(ruleset_name)
}
\arguments{
\item{ruleset_name}{Name of schema to get rules for}
}
\value{
Named list with 'stop_on_error' and 'collect_errors' rule functions
}
\description{
Get business rules for a given schema
}
