% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_connector_path.R
\name{get_data_connector_path}
\alias{get_data_connector_path}
\title{Get the path to the data connector for a given trial and data connection type}
\usage{
get_data_connector_path(
  type = c("sdtm", "adam", "metadata"),
  trial_metadata,
  data_connection = c("connector", "custom_data"),
  custom_data_path = NULL
)
}
\arguments{
\item{type}{A character string indicating the type of data connector. Valid
values are \code{type} (\code{sdtm}, \code{adam} or \code{metadata}.}

\item{trial_metadata}{A list containing trial metadata. It must contain the
\code{project_id}, \code{trial_id}, \code{complete_id}, and \code{instance}}

\item{data_connection}{A character string indicating the type of data
connection. Valid values are \code{connector} (returning a path in line with
current NN UNIX-SCE requirement), \code{custom_data} or \code{pharmaverse}}
}
\description{
This function generates the path to the data connector based on
the \code{type} (\code{sdtm}, \code{adam} or \code{metadata}),
\code{data_connection} type and the trial metadata.
}
\note{
This will become obsolete when connector config is used instead.
}
