% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_nodes.R
\name{group_actions}
\alias{group_actions}
\title{Group nodes based on domain}
\usage{
group_actions(ordered_nodes, vertices, edges)
}
\arguments{
\item{edges}{}
}
\description{
Group nodes based on domain
}
