% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_check_executable_status.R
\name{handle_filter_domain_action}
\alias{handle_filter_domain_action}
\title{Handler for _filter_domain.mustache}
\usage{
handle_filter_domain_action(action, processed_actions, available_cols, ui_data)
}
\description{
Handler for _filter_domain.mustache
}
