% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_check_executable_status.R
\name{handle_init_domain_action}
\alias{handle_init_domain_action}
\title{Handler for _init_domain.mustache actions}
\usage{
handle_init_domain_action(action, processed_actions, available_cols, ui_data)
}
\description{
Handler for _init_domain.mustache actions
}
