% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{handle_validation_errors}
\alias{handle_validation_errors}
\title{Handle validation errors with informative messages}
\usage{
handle_validation_errors(is_valid, yaml_file, verbose, error_handler = NULL)
}
\description{
Handle validation errors with informative messages
}
