% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_actions.R
\name{initialize_graph_data}
\alias{initialize_graph_data}
\title{Initialize all graph data structures needed for sorting}
\usage{
initialize_graph_data(edges, nodes, primary_domain)
}
\description{
Initialize all graph data structures needed for sorting
}
