% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_consolidated_env.R
\name{load_functions_from_files}
\alias{load_functions_from_files}
\title{Load Functions from Specified Source Files}
\usage{
load_functions_from_files(source_files, function_sources, envr)
}
\arguments{
\item{source_files}{Character vector specifying file paths of R scripts.}

\item{function_sources}{List used to track the origins of loaded functions,
including file sources.}

\item{envr}{Environment to which the loaded functions are assigned.}
}
\value{
The updated \code{function_sources$file} list containing details of
functions loaded from the specified files.
}
\description{
Loads functions from specified R script files into a given
environment.
}
\details{
For each given file, the function checks its existence (raising an
error if not found), sources it into a temporary environment, and assigns
any functions found to the specified environment \code{envr}. Function origins
are tracked within the \code{function_sources} list.
}
