% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_consolidated_env.R
\name{load_functions_from_packages}
\alias{load_functions_from_packages}
\title{Load Functions from Specified Packages}
\usage{
load_functions_from_packages(packages, code_ids, function_sources, envr)
}
\arguments{
\item{packages}{Character vector specifying the names of R packages to load
functions from.}

\item{code_ids}{Character vector containing the names of functions to be
loaded.}

\item{function_sources}{List used to track the origins of loaded functions,
including package sources.}

\item{envr}{Environment to which the loaded functions are assigned.}
}
\value{
The updated \code{function_sources$pkg} list containing details of
functions loaded from the specified packages.
}
\description{
Loads functions from specified R packages that match the given
\code{code_ids} and assigns them to a specified environment.
}
\details{
Iterates over the provided packages, verifies their availability,
and loads functions that match the provided \code{code_ids}. The functions are
assigned to the environment specified by \code{envr}, while their origins are
recorded in the \code{function_sources} list.
}
