% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_adam_program.R
\name{make_adam_program}
\alias{make_adam_program}
\title{make_adam_program}
\usage{
make_adam_program(path_ui_data, path_std_lib, path_output)
}
\arguments{
\item{path_output}{}
}
\description{
make_adam_program
}
