% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mighty-package.r
\docType{package}
\name{mighty-package}
\alias{mighty}
\alias{mighty-package}
\title{mighty: ADaM Program Generation Through Declarative Specifications}
\description{
Provides tools for generating ADaM programs in clinical trials using a declarative approach. Users specify what data structure and derivations they need through YAML-based specifications, rather than writing imperative code. The package handles dependency tracking between columns and rows, orchestrates code components, and automatically generates executable R programs for ADaM dataset creation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nn-opensource.github.io/mighty/}
}

}
\author{
\strong{Maintainer}: Matthew Phelps \email{mewp@novonordisk.com}

Authors:
\itemize{
  \item Nicolai Skov Johnsen \email{nosj@novonordisk.com}
  \item Aksel Thomsen \email{oath@novonordisk.com}
  \item Michael Hedegaard Thomsen \email{mhzt@novonordisk.com}
}

}
\keyword{internal}
