% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_init_domain_code.R
\name{params_init_domain_code}
\alias{params_init_domain_code}
\title{Write the R code to initialize ADaM domain in "initial" programs}
\usage{
params_init_domain_code(.self, keep_columns, source_domains)
}
\arguments{
\item{domain_filters_exist}{}
}
\description{
Write the R code to initialize ADaM domain in "initial" programs
}
