% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_code_components_metadata.R
\name{parse_code_components_metadata}
\alias{parse_code_components_metadata}
\title{Parses parent and children ADaM columns from functions' Roxygen header}
\usage{
parse_code_components_metadata(
  pkgs = NULL,
  source_files = NULL,
  function_names
)
}
\arguments{
\item{function_names}{}
}
\description{
The metadata regarding the ADaM columns each function requires to be
available to it are stored in the functions roxygen header, along with the
specification of which ADaM column(s) the function modifies/creates.
}
