% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_node_metadata.R
\name{parse_node_metadata}
\alias{parse_node_metadata}
\title{Get metadata from code components}
\usage{
parse_node_metadata(file_path)
}
\arguments{
\item{file_path}{}
}
\description{
Fetches code dependencies and outputs, as well as function
parameters (aside from the self and other data parameters)
}
