% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ui_data.R
\name{parse_ui_data}
\alias{parse_ui_data}
\title{Convienence function to parse the data from the UI}
\usage{
parse_ui_data(yml_data)
}
\description{
Performs various data integrity checks and actions. Ensurse NAs
are passed as R NA_character_ and not as strings of "na".
}
