% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_actions.R
\name{perform_topological_sort}
\alias{perform_topological_sort}
\title{Perform the main topological sorting algorithm}
\usage{
perform_topological_sort(graph_data, primary_domain)
}
\description{
Perform the main topological sorting algorithm
}
