% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_topo_order.R
\name{permute_topo_order}
\alias{permute_topo_order}
\title{Permute domain init nodes in topological order}
\usage{
permute_topo_order(topo_order_names, vertices)
}
\arguments{
\item{vertices}{}
}
\description{
Permute domain init nodes in topological order
}
