% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_predecessor_code.R
\name{pre_process_predecessor_left_join}
\alias{pre_process_predecessor_left_join}
\title{Pre-process variables for predecessor left joins}
\usage{
pre_process_predecessor_left_join(
  depend_columns,
  depend_domains,
  outputs,
  domain,
  domain_keys
)
}
\value{
a list of strings to be used in the predecessor_left_join function
}
\description{
Prepares the inputs to the predecessor_left_join function based
on the "nodes" data model
}
