% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_check_executable_status.R
\name{prefix_with_domain}
\alias{prefix_with_domain}
\title{Prefix output columns with \code{domain} name}
\usage{
prefix_with_domain(column, domain)
}
\description{
Prefix output columns with \code{domain} name
}
