% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_keys_to_depend_cols.R
\name{process_external_echo_dependencies}
\alias{process_external_echo_dependencies}
\title{Process External Dependencies for col_echo Actions}
\usage{
process_external_echo_dependencies(x, index_echos, pk)
}
\arguments{
\item{x}{A \code{data.table} of nodes representing an ADaM domain.}

\item{index_echos}{An integer vector of indices corresponding to col_echo
actions in the nodes.}

\item{pk}{A named list specifying primary key columns for external domains.}
}
\value{
A modified \code{data.table} of nodes with enriched dependencies for
col_echo actions.
}
\description{
Identifies and processes external dependencies in nodes for
col_echo actions.
}
\details{
Determines which col_echo actions reference external domains and
ensures that the appropriate foreign key columns are included in their
\code{depend_cols}. This step ensures that external dependencies are correctly
handled during processing.
}
