% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_domain_initialize_nodes.R
\name{replace_core_domain_with_adam_for_mutate_nodes}
\alias{replace_core_domain_with_adam_for_mutate_nodes}
\title{Replace "core" domains with ADaM "self" domain for col_mutate nodes}
\usage{
replace_core_domain_with_adam_for_mutate_nodes(nodes_subset, domain_init_data)
}
\arguments{
\item{domain_init_data}{}
}
\description{
Replace "core" domains with ADaM "self" domain for col_mutate nodes
}
\details{
col_mutate nodes are not absorbed by the domain_init nodes. However, the
dependencies specified in the col_mutate nodes should now point to the variables
outputted by the domain init node, not the original "core" domain(s). The
update is done by replacing the domain name in the depend_cols with the ADaM
domain name.
}
