% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_domain_initialize_nodes.R
\name{replace_core_domain_with_adam_for_rename_nodes}
\alias{replace_core_domain_with_adam_for_rename_nodes}
\title{Replace "core" domains with ADaM "self" domain for rename nodes}
\usage{
replace_core_domain_with_adam_for_rename_nodes(nodes_subset, domain_init_data)
}
\arguments{
\item{domain_init_data}{}
}
\description{
Replace "core" domains with ADaM "self" domain for rename nodes
}
\details{
Rename nodes are not absorbed by the domain_init nodes. However, the
dependencies specified in the rename nodes should now point to the variables
outputted by the domain init node, not the original "core" domain(s). The
update is done by replacing the domain name in the depend_cols with the ADaM
domain name.
}
