% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_self_with_domain.R
\name{replace_self_with_domain}
\alias{replace_self_with_domain}
\title{Add doamin information to any depends values that have ".self"}
\usage{
replace_self_with_domain(nodes)
}
\description{
This is needed so the dependencies between derivations in different tables
can be parsed
}
