% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{sc_extract_data_preview}
\alias{sc_extract_data_preview}
\title{Extract and format data preview from error}
\usage{
sc_extract_data_preview(
  error_row,
  format_dataframe_preview_fn,
  format_vector_preview_fn,
  format_atomic_preview_fn
)
}
\description{
Extract and format data preview from error
}
