% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{sc_format_additional_properties_error}
\alias{sc_format_additional_properties_error}
\title{Format additional properties errors}
\usage{
sc_format_additional_properties_error(
  error_row,
  extract_param_fn,
  format_path_fn
)
}
\description{
Format additional properties errors
}
