% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{sc_format_dataframe_preview}
\alias{sc_format_dataframe_preview}
\title{Format data frame preview for error messages}
\usage{
sc_format_dataframe_preview(data_content)
}
\description{
Format data frame preview for error messages
}
