% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{sc_format_errors}
\alias{sc_format_errors}
\title{Format validation errors with header and messages}
\usage{
sc_format_errors(errors, yaml_file, format_single_fn)
}
\description{
Format validation errors with header and messages
}
