% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{sc_format_generic_error}
\alias{sc_format_generic_error}
\title{Format generic validation errors with context}
\usage{
sc_format_generic_error(error_row, format_path_fn, extract_data_preview_fn)
}
\description{
Format generic validation errors with context
}
