% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{sc_format_min_items_error}
\alias{sc_format_min_items_error}
\title{Format minimum items errors}
\usage{
sc_format_min_items_error(error_row, extract_param_fn, format_path_fn)
}
\description{
Format minimum items errors
}
