% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{sc_format_pattern_error}
\alias{sc_format_pattern_error}
\title{Format pattern validation errors}
\usage{
sc_format_pattern_error(error_row, extract_param_fn, format_path_fn)
}
\description{
Format pattern validation errors
}
