% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{sc_format_single_error}
\alias{sc_format_single_error}
\title{Route error to appropriate formatter based on keyword}
\usage{
sc_format_single_error(error_row, formatters)
}
\description{
Route error to appropriate formatter based on keyword
}
