% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{sc_format_vector_preview}
\alias{sc_format_vector_preview}
\title{Format vector preview for error messages}
\usage{
sc_format_vector_preview(data_content)
}
\description{
Format vector preview for error messages
}
