% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{sc_setup_formatters}
\alias{sc_setup_formatters}
\title{Set up error formatters registry}
\usage{
sc_setup_formatters(extract_param_fn, format_path_fn, extract_data_preview_fn)
}
\description{
Set up error formatters registry
}
