% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_add_supp.R
\name{sdtm_add_supp}
\alias{sdtm_add_supp}
\title{Add supp data - our version of NNsdtm}
\usage{
sdtm_add_supp(data_main, data_supp, dataset = "dm")
}
\arguments{
\item{dataset}{}
}
\description{
This verison allows adding without a merging ID - but only for domains with 1 record per subject
}
