% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_actions.R
\name{select_node_by_priority}
\alias{select_node_by_priority}
\title{Select node based on priority rules (exact replication of original logic)}
\usage{
select_node_by_priority(
  candidate_nodes,
  sorted_nodes,
  adj_matrix,
  lkp_dom,
  primary_domain,
  domains_upstr_dep,
  domains_upstr_dep_prim_dom
)
}
\description{
Select node based on priority rules (exact replication of original logic)
}
