% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_actions.R
\name{sort_actions}
\alias{sort_actions}
\title{Sort actions according to topological order and dependencies}
\usage{
sort_actions(edges, nodes, primary_domain = "ADSL")
}
\arguments{
\item{primary_domain}{}
}
\description{
Sort actions according to topological order and dependencies
}
