% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_metadata.R
\name{table_metadata}
\alias{table_metadata}
\title{Table Metadata Class}
\description{
Stores metadata information for a single table including variable names and types.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{variables}}{Data.frame. Variables with name and type columns}

\item{\code{datasource}}{Character. Model type, i.e. \code{sdtm}, \code{adam} or \code{metadata}}

\item{\code{name}}{Character. Name of the table}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-table_metadata-new}{\code{table_metadata$new()}}
\item \href{#method-table_metadata-print}{\code{table_metadata$print()}}
\item \href{#method-table_metadata-clone}{\code{table_metadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-table_metadata-new"></a>}}
\if{latex}{\out{\hypertarget{method-table_metadata-new}{}}}
\subsection{Method \code{new()}}{
Create a new table_metadata object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{table_metadata$new(name, datasource, variables)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character. Table name}

\item{\code{datasource}}{Character. Model type (e.g. SDTM, ADAM, or METADATA)}

\item{\code{variables}}{Data.frame. Variable metadata}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-table_metadata-print"></a>}}
\if{latex}{\out{\hypertarget{method-table_metadata-print}{}}}
\subsection{Method \code{print()}}{
Print table metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{table_metadata$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-table_metadata-clone"></a>}}
\if{latex}{\out{\hypertarget{method-table_metadata-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{table_metadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
