% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_nodes.R
\name{traverse_and_group_nodes}
\alias{traverse_and_group_nodes}
\title{Title}
\usage{
traverse_and_group_nodes(
  n_remaining,
  n_stack,
  edges,
  node_group,
  node_id_i = NA,
  domain_i = NA,
  node_level_i = NA
)
}
\arguments{
\item{domain_i}{}
}
\description{
Title
}
