% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_check_executable_status.R
\name{update_available_columns_for_action}
\alias{update_available_columns_for_action}
\title{Update available columns after action execution}
\usage{
update_available_columns_for_action(action, available_cols)
}
\description{
Update available columns after action execution
}
