% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_depend_cols.R
\name{update_depend_cols}
\alias{update_depend_cols}
\title{Update update_depend_cols}
\usage{
update_depend_cols(nodes, pk, ui_init)
}
\arguments{
\item{ui_init}{}
}
\description{
Ensures that external domains

Additional details...
}
