% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_depend_rows.R
\name{update_depend_rows}
\alias{update_depend_rows}
\title{Update the depend_rows in a data.table by prefixing with the domain.}
\usage{
update_depend_rows(nodes)
}
\arguments{
\item{nodes}{A data.table with at least two columns:
\describe{
\item{depend_rows}{A list or character vector containing dependent rows.}
\item{domain}{A character vector used as a prefix.}
}}
}
\value{
A data.table with updated \code{depend_rows} where prefixes are added.
}
\description{
This function takes a data.table containing a column \code{depend_rows} and
a column \code{domain}. It updates the \code{depend_rows} by prepending the
corresponding \code{domain} value to each element in \code{depend_rows} for rows
where \code{depend_rows} is not NA.
}
