% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_predecessors.R
\name{update_predecessors}
\alias{update_predecessors}
\title{Update predecessors}
\usage{
update_predecessors(nodes, path_mappings)
}
\arguments{
\item{path_mappings}{}
}
\description{
Ensures that external domains

Additional details...
}
