% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{validate_json_against_schema}
\alias{validate_json_against_schema}
\title{Validate JSON string against schema}
\usage{
validate_json_against_schema(
  json_data,
  schema_path,
  yaml_file,
  verbose = TRUE,
  error_handler = NULL
)
}
\description{
Validate JSON string against schema
}
