% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{ValidationErrorHandler}
\alias{ValidationErrorHandler}
\title{Validation Error Handler Class}
\description{
Encapsulates error formatting logic with extensible formatter registry
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ValidationErrorHandler-new}{\code{ValidationErrorHandler$new()}}
\item \href{#method-ValidationErrorHandler-format_errors}{\code{ValidationErrorHandler$format_errors()}}
\item \href{#method-ValidationErrorHandler-format_single_error}{\code{ValidationErrorHandler$format_single_error()}}
\item \href{#method-ValidationErrorHandler-clone}{\code{ValidationErrorHandler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorHandler-new"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorHandler-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorHandler$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorHandler-format_errors"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorHandler-format_errors}{}}}
\subsection{Method \code{format_errors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorHandler$format_errors(errors, yaml_file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorHandler-format_single_error"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorHandler-format_single_error}{}}}
\subsection{Method \code{format_single_error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorHandler$format_single_error(error_row)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorHandler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorHandler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
