% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_node_topo_sort.R
\name{weighted_node_topo_sort}
\alias{weighted_node_topo_sort}
\title{Topological sort of a weighted directed graph}
\usage{
weighted_node_topo_sort(edges, nodes, primary_domain = "ADSL")
}
\arguments{
\item{primary_domain}{}
}
\description{
Topological sort of a weighted directed graph
}
