% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_adam_domain_yml.R
\name{write_adam_domain_yml}
\alias{write_adam_domain_yml}
\title{Write Single ADaM Domain Specification to YAML File}
\usage{
write_adam_domain_yml(spec, output_path)
}
\arguments{
\item{spec}{A single ADAM domain specification object (list) containing
the internal representation of column metadata, row actions, and other
domain-specific information.}

\item{output_path}{Character string specifying the full path for the output
YAML file, including filename and extension.}
}
\value{
Character string of the output file path.
}
\description{
Writes a single ADaM domain specification to a YAML file.
}
