% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_adam_programs.R
\name{write_adam_programs}
\alias{write_adam_programs}
\title{Write the ADaM progrms to disk and style them}
\usage{
write_adam_programs(programs, dir)
}
\arguments{
\item{dir}{}
}
\description{
Write the ADaM progrms to disk and style them
}
