% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_actions.R
\name{apply_priority_2}
\alias{apply_priority_2}
\title{Apply Priority 2: Complex domain dependency logic}
\usage{
apply_priority_2(
  candidate_nodes,
  adj_matrix,
  lkp_dom,
  primary_domain,
  domains_upstr_dep,
  domains_upstr_dep_prim_dom
)
}
\description{
Apply Priority 2: Complex domain dependency logic
}
