% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_actions.R
\name{apply_priority_2_subpriorities}
\alias{apply_priority_2_subpriorities}
\title{Apply Priority 2 sub-priorities for covered candidates}
\usage{
apply_priority_2_subpriorities(
  candidate_nodes_upstr_dep_cov,
  adj_matrix,
  lkp_dom,
  primary_domain,
  domains_upstr_dep,
  domains_upstr_dep_prim_dom
)
}
\description{
Apply Priority 2 sub-priorities for covered candidates
}
