% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_outputs_identical.R
\name{assert_outputs_identical}
\alias{assert_outputs_identical}
\title{Assert outputs defined in code components match those defined in yml specs}
\usage{
assert_outputs_identical(x)
}
\arguments{
\item{x}{}
}
\value{
TRUE if passes, error if it does not
}
\description{
Assert outputs defined in code components match those defined in yml specs
}
