% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_yaml.R
\name{convert_yaml_to_json}
\alias{convert_yaml_to_json}
\title{Convert YAML file to JSON string using appropriate method}
\usage{
convert_yaml_to_json(yaml_file, use_yq = TRUE, verbose = TRUE)
}
\description{
Convert YAML file to JSON string using appropriate method
}
