% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_yaml.R
\name{convert_yaml_to_json_with_r}
\alias{convert_yaml_to_json_with_r}
\title{Convert YAML to JSON using R yaml package}
\usage{
convert_yaml_to_json_with_r(yaml_file)
}
\description{
Convert YAML to JSON using R yaml package
}
