% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_check_executable_status.R
\name{find_missing_columns}
\alias{find_missing_columns}
\title{Find columns that are missing from available columns}
\usage{
find_missing_columns(needed_outputs, available_cols)
}
\description{
Find columns that are missing from available columns
}
