% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_check_executable_status.R
\name{find_missing_dependencies}
\alias{find_missing_dependencies}
\title{Find missing dependencies by comparing depend_cols with available columns}
\usage{
find_missing_dependencies(depend_cols, available_cols)
}
\description{
Find missing dependencies by comparing depend_cols with available columns
}
