% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adam_code.R
\name{generate_adam_code}
\alias{generate_adam_code}
\title{Generates the complete set of ADaM programs}
\usage{
generate_adam_code(
  path_ui_data,
  code_component_source_pkgs = NULL,
  code_component_source_files = NULL,
  path_trial_metadata,
  path_domain_keys,
  path_output,
  data_connection = c("connector", "pharmaverse", "custom_data"),
  connector_config_path = NULL,
  check_cross_domain_adam_dependencies = TRUE
)
}
\arguments{
\item{check_cross_domain_adam_dependencies}{}
}
\description{
Generates the complete set of ADaM programs
}
