% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adam_code.R
\name{generate_adam_code}
\alias{generate_adam_code}
\title{Generates the complete set of ADaM programs}
\usage{
generate_adam_code(
  path_ui_data,
  standards_lib = NULL,
  path_trial_metadata,
  path_trial,
  check_cross_domain_adam_dependencies = TRUE,
  data_context = NULL
)
}
\arguments{
\item{path_trial_metadata}{}
}
\description{
Generates the complete set of ADaM programs
}
