% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_external_data_code.R
\name{generate_external_data_code}
\alias{generate_external_data_code}
\title{Generate external data code}
\usage{
generate_external_data_code(
  payload,
  trial_metadata,
  sdtm_dataset_list,
  adam_dataset_list,
  data_connection,
  path_output
)
}
\arguments{
\item{custom_data_path}{}
}
\description{
Generate external data code
}
