% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_initialize_domain.R
\name{generate_initialize_domain}
\alias{generate_initialize_domain}
\title{Write the R code to initialize a domain}
\usage{
generate_initialize_domain(
  .self,
  core_domains,
  filter_per_domain = NULL,
  filter_global = NULL,
  keep_vars = NULL
)
}
\arguments{
\item{.self}{Character of length 1. Name of domain}

\item{core_domains}{list of core SDTM domains used to make "base" table}

\item{filter_global}{String. Global filter to be applied after merging}

\item{keep_vars}{Vector of strings naming variables from SDTM core domains
kept}

\item{filter_expr}{list of stings with lenght equal to number of core_domains
filters to apply to each core SDTM table before combining}
}
\description{
Write the R code to initialize a domain
}
