% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_program_init.R
\name{generate_program_init}
\alias{generate_program_init}
\title{Write the R code to initialize an ADaM "update" program}
\usage{
generate_program_init(adam_domain, adam_dataset_list, path_out)
}
\arguments{
\item{adam_dataset_list}{}
}
\description{
Write the R code to initialize an ADaM "update" program
}
