% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_read_data_code.R
\name{generate_read_data_code}
\alias{generate_read_data_code}
\title{Generate external data code}
\usage{
generate_read_data_code(
  payload,
  trial_metadata,
  sdtm_dataset_list,
<<<<<<< HEAD:man/generate_read_data_code.Rd
  data_connection,
  path_output,
  core_domains,
  .self,
  domain_filters_exist
=======
  path_output
>>>>>>> main:man/generate_external_data_code.Rd
)
}
\arguments{
\item{path_output}{}
}
\description{
Generate external data code
}
