% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_read_data_code.R
\name{generate_read_data_code}
\alias{generate_read_data_code}
\title{Generate external data code}
\usage{
generate_read_data_code(payload, sdtm_dataset_list, path_output)
}
\arguments{
\item{data_connection}{}
}
\description{
Generate external data code
}
