% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_check_executable_status.R
\name{handle_generic_action}
\alias{handle_generic_action}
\title{Handler for generic actions (default case)}
\usage{
handle_generic_action(action, processed_actions, available_cols, ui_data)
}
\description{
Handler for generic actions (default case)
}
