% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_check_executable_status.R
\name{handle_read_data_action}
\alias{handle_read_data_action}
\title{Handler for _read_data.mustache actions}
\usage{
handle_read_data_action(action, processed_actions, available_cols, ui_data)
}
\description{
Handler for _read_data.mustache actions
}
