% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_adam_program.R
\name{make_adam_program}
\alias{make_adam_program}
\title{make_adam_program}
\usage{
make_adam_program(
  path_ui_data,
  path_std_lib,
  path_domain_keys,
  path_output,
  data_connection = c("connector", "pharmaverse")
)
}
\arguments{
\item{data_connection}{}
}
\description{
make_adam_program
}
