% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mighty-package.r
\docType{package}
\name{mighty-package}
\alias{mighty}
\alias{mighty-package}
\title{mighty: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Matthew Phelps \email{mewp@novonordisk.com}

Authors:
\itemize{
  \item Nicolai Skov Johnsen \email{nosj@novonordisk.com}
  \item Aksel Thomsen \email{oath@novonordisk.com}
  \item Michael Hedegaard Thomsen \email{mhzt@novonordisk.com}
}

}
\keyword{internal}
