% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_read_data_code.R
\name{params_read_data_code}
\alias{params_read_data_code}
\title{Generate external data code}
\usage{
params_read_data_code(payload, domain, path_trial)
}
\arguments{
\item{path_trial}{}
}
\description{
Generate external data code
}
