% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_into_chunks.R
\name{parse_into_chunks}
\alias{parse_into_chunks}
\title{Parse into chunks}
\usage{
parse_into_chunks(code_id, action, domain_name, env, action_name)
}
\arguments{
\item{action_name}{}
}
\description{
Parse into chunks
}
