% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_check_executable_status.R
\name{parse_output_columns}
\alias{parse_output_columns}
\title{Parse output columns from dot-separated format}
\usage{
parse_output_columns(outputs, default_prefix = "")
}
\description{
Parse output columns from dot-separated format
}
