% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_check_executable_status.R
\name{process_action}
\alias{process_action}
\title{Process a single action based on its type}
\usage{
process_action(action, processed_actions, available_cols, ui_data = NULL)
}
\description{
Process a single action based on its type
}
