% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_actions.R
\name{remove_selected_nodes}
\alias{remove_selected_nodes}
\title{Remove selected nodes from adjacency matrix}
\usage{
remove_selected_nodes(adj_matrix, selected_nodes)
}
\description{
Remove selected nodes from adjacency matrix
}
